requirements.txt
streamlit
openai
pdfplumber
pandas
python-dotenv
fastapi
uvicorn
gunicorn
python-multipart
